Consolidate rustdoc's lint passes into a single pass
This should improve performance and simplify the code.
